requirements.txt
numpy
scikit-learn
umap-learn